Texture Optimisation
Textures have been optimised for reducing shader amount.
The original version was optimised so that there would be a shader per texture, per shadertype. Which in theory is good if there's lot's of objects that uses one texture. However I'm going under the assumption that the game will not have a performance benefit from that, allowing me to trade textures at runtime for a single shader.